Refactor de modulos de tests
refs:#87
