+++ tests/support/factories/catalog_errors.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-from .utils import jsonschema_str
+from tests.support.utils import jsonschema_str
 
 
-def catalog_error_response(options=None):
+def catalog_error(options=None):
     default_options = {
         'title': "Datos Argentina",
         'message': None,
@@ -56,27 +56,27 @@ def catalog_error_response(options=None):
 
 
 def missing_catalog_dataset():
-    return catalog_error_response({
+    return catalog_error({
         'message': "%s is a required property" % jsonschema_str('dataset'),
         'dataset': None,
     })
 
 
 def missing_catalog_title():
-    return catalog_error_response({
+    return catalog_error({
         'message': "%s is a required property" % jsonschema_str('title'),
         'title': None,
     })
 
 
 def missing_catalog_description():
-    return catalog_error_response({
+    return catalog_error({
         'message': "%s is a required property" % jsonschema_str('description'),
     })
 
 
 def invalid_catalog_publisher_type():
-    return catalog_error_response({
+    return catalog_error({
         'instance': [
             "Ministerio de Modernización",
             "datos@modernizacion.gob.ar"
@@ -95,7 +95,7 @@ def invalid_catalog_publisher_type():
 
 
 def invalid_publisher_mbox_format():
-    return catalog_error_response({
+    return catalog_error({
         "error_code": 2,
         "message": "%s is not a %s" % (
             jsonschema_str('datosATmodernizacion.gob.ar'), jsonschema_str('email')),
@@ -110,7 +110,7 @@ def invalid_publisher_mbox_format():
 
 
 def null_catalog_publisher():
-    return catalog_error_response({
+    return catalog_error({
         "error_code": 2,
         "message": "None is not of type %s" % jsonschema_str('object'),
         "path": ['publisher'],
@@ -120,7 +120,7 @@ def null_catalog_publisher():
 
 
 def empty_mandatory_string():
-    return catalog_error_response({
+    return catalog_error({
         "error_code": 2,
         "message": "%s is too short" % jsonschema_str(''),
         "path": ['description'],
@@ -131,7 +131,7 @@ def empty_mandatory_string():
 
 
 def malformed_date():
-    return catalog_error_response({
+    return catalog_error({
         "instance": "2016/04/14",
         "validator": "anyOf",
         "path": [
@@ -160,7 +160,7 @@ def malformed_date():
 
 
 def malformed_datetime():
-    return catalog_error_response({
+    return catalog_error({
         "instance": "2016-04-1419:48:05.433640-03:00",
         "validator": "anyOf",
         "path": [
@@ -190,7 +190,7 @@ def malformed_datetime():
 
 
 def malformed_datetime2():
-    return catalog_error_response({
+    return catalog_error({
         "instance": "2016-04-54T19:48:05.433640-03:00",
         "validator": "anyOf",
         "path": [
@@ -220,7 +220,7 @@ def malformed_datetime2():
 
 
 def malformed_email():
-    return catalog_error_response({
+    return catalog_error({
         "instance": "datosATmodernizacion.gob.ar",
         "validator": "format",
         "path": [
@@ -235,7 +235,7 @@ def malformed_email():
 
 
 def malformed_uri():
-    return catalog_error_response({
+    return catalog_error({
         "instance": "datos.gob.ar/superThemeTaxonomy.json",
         "validator": "format",
         "path": [
@@ -249,7 +249,7 @@ def malformed_uri():
 
 
 def invalid_theme_taxonomy():
-    return catalog_error_response({
+    return catalog_error({
         "instance": None,
         "validator": "repeatedValue",
         "path": [
@@ -264,7 +264,7 @@ def invalid_theme_taxonomy():
 
 
 def missing_dataset():
-    return catalog_error_response({
+    return catalog_error({
         "instance": None,
         "validator": "required",
         "path": [],
