+++ tests/test_readers_and_writers.py
@@ -11,7 +11,7 @@ import unittest
 import nose
 import vcr
 
-from .factories.xlsx import CSV_TABLE, WRITE_XLSX_TABLE, READ_XLSX_TABLE
+from tests.support.factories.xlsx import CSV_TABLE, WRITE_XLSX_TABLE, READ_XLSX_TABLE
 
 try:
     import mock
